+++ lazarkit/src/main/web/src/lib/ui/primitives/index.ts
@@ -1,0 +1,4 @@
+import Section from "./Section.svelte"
+import Button from "./Button.svelte"
+
+export { Section, Button }
