+++ lazarkit/src/main/web/src/lib/ui/widgets/index.ts
@@ -1,0 +1,4 @@
+import OpModeControl from "./OpModeControl.svelte"
+import Telemetry from "./Telemetry.svelte"
+import Configurables from "./Configurables.svelte"
+export { OpModeControl, Telemetry, Configurables }
